move auth to linked_domain app
