+++ corehq/apps/linked_domain/auth.py
@@ -1,0 +1,20 @@
+from requests.auth import AuthBase
+
+
+class ApiKeyAuth(AuthBase):
+    def __init__(self, username, apikey):
+        self.username = username
+        self.apikey = apikey
+
+    def __eq__(self, other):
+        return all([
+            self.username == getattr(other, 'username', None),
+            self.apikey == getattr(other, 'apikey', None)
+        ])
+
+    def __ne__(self, other):
+        return not self == other
+
+    def __call__(self, r):
+        r.headers['Authorization'] = 'apikey %s:%s' % (self.username, self.apikey)
+        return r
